fix: resolve persona prompt from live record at spawn for non-pack agents
Persona edits never reached agents created from a plain persona (not an
installed pack): spawn only consulted the live PersonaRecord when both
persona_pack_path and persona_name_in_pack were set, so non-pack agents
fell through to the frozen record.system_prompt snapshot. Editing the
persona left that snapshot stale, and no restart could pick it up.

Resolve the linked persona whenever persona_id is set (pack or not); the
persona is the source of truth and edits propagate on the next spawn.
Fall back to the record snapshot only when no persona is linked or it was
deleted. Pack agents are unaffected (they already set persona_id).

Extract the resolution into a pure helper and move the runtime test module
to runtime/tests.rs (env_vars/tests.rs pattern), keeping runtime.rs under
the file-size limit. 4 new unit tests cover persona-wins, no-persona
fallback, deleted-persona fallback, and persona-clears-stale-model.
