+++ desktop/src-tauri/src/managed_agents/runtime.rs
@@ -6,7 +6,7 @@ use crate::{
     managed_agents::{
         append_log_marker, known_acp_runtime, login_shell_path, managed_agent_log_path,
         missing_command_message, normalize_agent_args, open_log_file, resolve_command,
-        ManagedAgentProcess, ManagedAgentRecord, ManagedAgentSummary, PersonaRecord,
+        ManagedAgentProcess, ManagedAgentRecord, ManagedAgentSummary,
     },
     util::now_iso,
 };
@@ -768,6 +768,26 @@ pub(crate) fn build_respond_to_env(
     Ok((set, remove))
 }
 
+/// Resolve the effective system prompt, model, and provider for a spawn. The
+/// linked persona always wins so persona edits propagate on the next spawn; the
+/// record snapshot is the fallback only when no persona is linked or it was
+/// deleted. Provider comes from the persona (the record has no provider field).
+fn resolve_effective_prompt_model_provider(
+    persona_id: Option<&str>,
+    personas: &[crate::managed_agents::types::PersonaRecord],
+    record_prompt: Option<String>,
+    record_model: Option<String>,
+) -> (Option<String>, Option<String>, Option<String>) {
+    match persona_id.and_then(|pid| personas.iter().find(|p| p.id == pid)) {
+        Some(p) => (
+            Some(p.system_prompt.clone()),
+            p.model.clone(),
+            p.provider.clone(),
+        ),
+        None => (record_prompt, record_model, None),
+    }
+}
+
 /// Spawn an agent process without holding any locks on records or runtimes.
 /// Returns the child process and log path on success. The caller is responsible
 /// for updating `ManagedAgentRecord` fields and inserting into the runtimes map.
@@ -900,32 +920,18 @@ pub fn spawn_agent_child(
         command.env("SPROUT_ACP_PERSONA_NAME", persona_name);
     }
 
-    // Resolve system prompt, model, and provider: prefer the persona definition
-    // (if a persona pack is configured and the persona matched), otherwise fall
-    // back to the record-level overrides. Provider always flows from the persona
-    // when one is linked (the record has no provider field of its own).
-    let has_persona_pack =
-        record.persona_pack_path.is_some() && record.persona_name_in_pack.is_some();
-    let persona_record: Option<PersonaRecord> = record.persona_id.as_deref().and_then(|pid| {
-        super::load_personas(app)
-            .ok()?
-            .into_iter()
-            .find(|p| p.id == pid)
-    });
-
-    let (effective_prompt, effective_model, effective_provider) = if has_persona_pack {
-        match &persona_record {
-            Some(p) => (
-                Some(p.system_prompt.clone()),
-                p.model.clone(),
-                p.provider.clone(),
-            ),
-            None => (record.system_prompt.clone(), record.model.clone(), None),
-        }
-    } else {
-        let provider = persona_record.as_ref().and_then(|p| p.provider.clone());
-        (record.system_prompt.clone(), record.model.clone(), provider)
-    };
+    // Resolve system prompt, model, and provider: the linked persona is the
+    // source of truth, so persona edits reach the agent on the next spawn. Fall
+    // back to the record snapshot only when no persona is linked or it was
+    // deleted. Provider flows from the persona (the record has no provider).
+    let personas = super::load_personas(app).unwrap_or_default();
+    let (effective_prompt, effective_model, effective_provider) =
+        resolve_effective_prompt_model_provider(
+            record.persona_id.as_deref(),
+            &personas,
+            record.system_prompt.clone(),
+            record.model.clone(),
+        );
 
     if let Some(prompt) = &effective_prompt {
         command.env("SPROUT_ACP_SYSTEM_PROMPT", prompt);
@@ -1219,244 +1225,4 @@ fn runtime_metadata_env_vars<'a>(
 }
 
 #[cfg(test)]
-mod tests {
-    use crate::managed_agents::known_acp_runtime;
-
-    #[test]
-    fn marker_entry_is_namespaced_by_instance_id() {
-        // The spawn stamp and the sweep matcher must produce identical bytes;
-        // both go through sprout_marker_entry, so this pins the on-the-wire
-        // format and guards against a dev build (`...app.dev`) matching a
-        // release build's (`...app`) agents.
-        assert_eq!(
-            super::sprout_marker_entry("xyz.block.sprout.app"),
-            b"SPROUT_MANAGED_AGENT=xyz.block.sprout.app".to_vec()
-        );
-        assert_ne!(
-            super::sprout_marker_entry("xyz.block.sprout.app"),
-            super::sprout_marker_entry("xyz.block.sprout.app.dev")
-        );
-    }
-
-    #[test]
-    fn sprout_agent_has_mcp_hooks() {
-        let p = known_acp_runtime("sprout-agent").expect("should resolve");
-        assert!(p.mcp_hooks);
-        assert_eq!(p.mcp_command, Some("sprout-dev-mcp"));
-    }
-
-    #[test]
-    fn databricks_defaults_empty_in_oss_build() {
-        // OSS (and normal test) builds set neither SPROUT_BUILD_DATABRICKS_*,
-        // so nothing is baked in and no DATABRICKS_* is injected on spawn.
-        assert!(super::build_databricks_defaults().is_empty());
-    }
-
-    #[test]
-    fn sprout_agent_resolved_via_path() {
-        assert!(known_acp_runtime("/usr/local/bin/sprout-agent").is_some_and(|p| p.mcp_hooks));
-    }
-
-    #[test]
-    fn goose_has_no_mcp_hooks() {
-        let p = known_acp_runtime("goose").expect("should resolve");
-        assert!(!p.mcp_hooks);
-        assert_eq!(p.mcp_command, None);
-    }
-
-    #[test]
-    fn unknown_command_returns_none() {
-        assert!(known_acp_runtime("custom-agent").is_none());
-    }
-
-    // ── build_respond_to_env tests ───────────────────────────────────────
-
-    use super::build_respond_to_env;
-    use crate::managed_agents::types::{ManagedAgentRecord, RespondTo};
-
-    /// Construct a minimal record fixture for env-building tests. Only the
-    /// fields read by `build_respond_to_env` matter here.
-    fn fixture(
-        respond_to: RespondTo,
-        allowlist: Vec<String>,
-        auth_tag: Option<String>,
-    ) -> ManagedAgentRecord {
-        ManagedAgentRecord {
-            pubkey: "p".into(),
-            name: "n".into(),
-            persona_id: None,
-            private_key_nsec: "nsec1fake".into(),
-            auth_tag,
-            relay_url: "ws://localhost:3000".into(),
-            avatar_url: None,
-            acp_command: "sprout-acp".into(),
-            agent_command: "goose".into(),
-            agent_args: vec![],
-            mcp_command: String::new(),
-            turn_timeout_seconds: 320,
-            idle_timeout_seconds: None,
-            max_turn_duration_seconds: None,
-            parallelism: 1,
-            system_prompt: None,
-            model: None,
-            mcp_toolsets: None,
-            env_vars: std::collections::BTreeMap::new(),
-            start_on_app_launch: false,
-            runtime_pid: None,
-            backend: Default::default(),
-            backend_agent_id: None,
-            provider_binary_path: None,
-            persona_pack_path: None,
-            persona_name_in_pack: None,
-            created_at: "now".into(),
-            updated_at: "now".into(),
-            last_started_at: None,
-            last_stopped_at: None,
-            last_exit_code: None,
-            last_error: None,
-            respond_to,
-            respond_to_allowlist: allowlist,
-            relay_mesh: None,
-        }
-    }
-
-    #[test]
-    fn build_env_owner_only_sets_mode_and_removes_others() {
-        let rec = fixture(RespondTo::OwnerOnly, vec![], Some("tag".into()));
-        let (set, remove) = build_respond_to_env(&rec, Some("owner")).unwrap();
-        let set_map: std::collections::HashMap<_, _> = set.into_iter().collect();
-        assert_eq!(
-            set_map.get("SPROUT_ACP_RESPOND_TO").map(String::as_str),
-            Some("owner-only")
-        );
-        assert!(!set_map.contains_key("SPROUT_ACP_RESPOND_TO_ALLOWLIST"));
-        assert!(remove.contains(&"SPROUT_ACP_RESPOND_TO_ALLOWLIST"));
-        // auth_tag is present → no AGENT_OWNER fallback fires.
-        assert!(remove.contains(&"SPROUT_ACP_AGENT_OWNER"));
-    }
-
-    #[test]
-    fn build_env_allowlist_sets_both_envs_and_joins() {
-        let a = "a".repeat(64);
-        let b = "b".repeat(64);
-        let rec = fixture(
-            RespondTo::Allowlist,
-            vec![a.clone(), b.clone()],
-            Some("tag".into()),
-        );
-        let (set, _remove) = build_respond_to_env(&rec, Some("owner")).unwrap();
-        let set_map: std::collections::HashMap<_, _> = set.into_iter().collect();
-        assert_eq!(
-            set_map.get("SPROUT_ACP_RESPOND_TO").map(String::as_str),
-            Some("allowlist")
-        );
-        assert_eq!(
-            set_map
-                .get("SPROUT_ACP_RESPOND_TO_ALLOWLIST")
-                .map(String::as_str),
-            Some(format!("{a},{b}").as_str()),
-        );
-    }
-
-    #[test]
-    fn build_env_anyone_omits_allowlist_var() {
-        let rec = fixture(RespondTo::Anyone, vec![], Some("tag".into()));
-        let (set, remove) = build_respond_to_env(&rec, Some("owner")).unwrap();
-        let set_map: std::collections::HashMap<_, _> = set.into_iter().collect();
-        assert_eq!(
-            set_map.get("SPROUT_ACP_RESPOND_TO").map(String::as_str),
-            Some("anyone")
-        );
-        assert!(!set_map.contains_key("SPROUT_ACP_RESPOND_TO_ALLOWLIST"));
-        assert!(remove.contains(&"SPROUT_ACP_RESPOND_TO_ALLOWLIST"));
-    }
-
-    #[test]
-    fn build_env_legacy_record_without_auth_tag_emits_agent_owner() {
-        let rec = fixture(RespondTo::OwnerOnly, vec![], None);
-        let (set, remove) = build_respond_to_env(&rec, Some("ownerhex")).unwrap();
-        let set_map: std::collections::HashMap<_, _> = set.into_iter().collect();
-        assert_eq!(
-            set_map.get("SPROUT_ACP_AGENT_OWNER").map(String::as_str),
-            Some("ownerhex")
-        );
-        assert!(!remove.contains(&"SPROUT_ACP_AGENT_OWNER"));
-    }
-
-    #[test]
-    fn build_env_legacy_record_without_owner_hex_removes_agent_owner() {
-        // No owner available to forward → make sure we don't inherit a leaked
-        // env var from the parent.
-        let rec = fixture(RespondTo::OwnerOnly, vec![], None);
-        let (_set, remove) = build_respond_to_env(&rec, None).unwrap();
-        assert!(remove.contains(&"SPROUT_ACP_AGENT_OWNER"));
-    }
-
-    #[test]
-    fn build_env_rejects_corrupted_allowlist() {
-        let rec = fixture(
-            RespondTo::Allowlist,
-            vec!["not-hex".into()],
-            Some("tag".into()),
-        );
-        assert!(build_respond_to_env(&rec, Some("owner")).is_err());
-    }
-
-    #[test]
-    fn build_env_rejects_empty_allowlist_in_allowlist_mode() {
-        let rec = fixture(RespondTo::Allowlist, vec![], Some("tag".into()));
-        let err = build_respond_to_env(&rec, Some("owner")).unwrap_err();
-        assert!(err.contains("at least one pubkey"));
-    }
-
-    // ── runtime_metadata_env_vars tests ─────────────────────────────────────
-
-    use super::runtime_metadata_env_vars;
-
-    #[test]
-    fn runtime_metadata_env_vars_injects_model_and_provider() {
-        let vars = runtime_metadata_env_vars(
-            Some("GOOSE_MODEL"),
-            Some("GOOSE_PROVIDER"),
-            false,
-            Some("gpt-4o"),
-            Some("openai"),
-        );
-        assert_eq!(
-            vars,
-            vec![("GOOSE_MODEL", "gpt-4o"), ("GOOSE_PROVIDER", "openai")]
-        );
-    }
-
-    #[test]
-    fn runtime_metadata_env_vars_skips_provider_when_locked() {
-        let vars = runtime_metadata_env_vars(
-            None, // claude has no model_env_var
-            None, // claude has no provider_env_var
-            true, // provider_locked = true
-            Some("claude-opus-4-7"),
-            Some("anthropic"),
-        );
-        assert!(vars.is_empty());
-    }
-
-    #[test]
-    fn runtime_metadata_env_vars_injects_model_even_with_acp_model_switching() {
-        // sprout-agent has supports_acp_model_switching=true but we still inject
-        // the model env var because ACP model switching is post-bootstrap
-        let vars = runtime_metadata_env_vars(
-            Some("SPROUT_AGENT_MODEL"),
-            Some("SPROUT_AGENT_PROVIDER"),
-            false,
-            Some("goose-claude-4-6-opus"),
-            Some("databricks"),
-        );
-        assert_eq!(
-            vars,
-            vec![
-                ("SPROUT_AGENT_MODEL", "goose-claude-4-6-opus"),
-                ("SPROUT_AGENT_PROVIDER", "databricks"),
-            ]
-        );
-    }
-}
+mod tests;
